Fix issue where yield_to can return too early
If the signal emits the value 'false' as the first argument this will
assume that it's the bound 'false' value which signifies a timeout..
This is because bound arguments appear at the end of the callback
parameter list, not a the beginning.

Instead of using true/false, this change binds a sentinel. We can
definitively know whether the callback came from the bound signal or a
timeout.

See also https://github.com/bitwes/Gut/issues/327